Renamed and refactored some objects to use lombok
